Fix validation of MIME types for files (#2445)
## Description

- **Summary of changes**: use a Literal list of allowed MIME types
instead of a class attribute, which doesn't work
- **Related issues**: Following up on
https://github.com/agno-agi/agno/pull/2325, that introduced file
handling for Claude and Gemini

The code currently fails with invalid attribute for
`cls.VALID_MIME_TYPES`.

The simplest solution is to use a Literal type and let pydantic validate
the value.

Potential but not concrete drawbacks:
* The exception raised changes from ValueError to ValidationError;
* In case someone wants to know which MIME types are supported, they'll
have to use something like `typing.get_args(File.mime_type)` instead of
`File.VALID_MIME_TYPES`. However, it should be noted that right now this
list is not truly reflecting the supported types for a model (example:
Claude only supports `application/pdf`).

Fixes #(issue)

---

## Type of change

Please check the options that are relevant:

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Model update (Addition or modification of models)
- [ ] Other (please describe):

---

## Checklist

- [ ] Adherence to standards: Code complies with Agno’s style guidelines
and best practices.
- [x] Formatting and validation: You have run `./scripts/format.sh` and
`./scripts/validate.sh` to ensure code is formatted and linted.
- [x] Self-review completed: A thorough review has been performed by the
contributor(s).
- [ ] Documentation: Docstrings and comments have been added or updated
for any complex logic.
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable).
- [x] Tested in a clean environment: Changes have been tested in a clean
environment to confirm expected behavior.
- [ ] Tests (optional): Tests have been added or updated to cover any
new or changed functionality.