Add fallback to deprecated imghdr (#2487)
## Description

- **Summary of changes**: 
Add fallback logic to the use of `imghdr` when this package is not
available.
Add support for 'jpg' types (simply add an entry to the type mapping as
`.jpg` and `.jpeg` files are handled equally in Claude
- **Motivation and context**:
The `imghdr` has been deprecated since Python3.11 (see here:
https://docs.python.org/3/library/imghdr.html).
The `claude` model of Agno makes use of this package. and this MR aims
to add a solution for the more recent versions of Python while keeping
backwards compatibility.
- **Environment or dependencies**: Specify any changes in dependencies
or environment configurations required for this update.
The package `filetype` must be installed by the user if he wishes to add
image support when using Claude


## Type of change

Please check the options that are relevant:

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Model update (Addition or modification of models)
- [ ] Other (please describe):

---

## Checklist

- [x] Adherence to standards: Code complies with Agno’s style guidelines
and best practices.
- [x] Formatting and validation: You have run `./scripts/format.sh` and
`./scripts/validate.sh` to ensure code is formatted and linted.
- [x] Self-review completed: A thorough review has been performed by the
contributor(s).
- [x] Documentation: Docstrings and comments have been added or updated
for any complex logic.
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable).
- [x] Tested in a clean environment: Changes have been tested in a clean
environment to confirm expected behavior.
- [ ] Tests (optional): Tests have been added or updated to cover any
new or changed functionality.

---------

Co-authored-by: Sebastião Assunção <sebastiao.assuncao@unbabel.com>
Co-authored-by: Dirk Brand <dirkbrnd@gmail.com>