Extracting website main content should return content as text if all else fails (#2804)
## Summary

This fixes the problem with website reader where if the content does not
have "content" tag the result would be empty. We want instead to return
the page as text so that we return at least something.

## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [x] Code complies with style guidelines
- [x] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [x] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [ ] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

This should not break any existing code.

Co-authored-by: Martin Schröder <info@swedishembedded.com>