fix(evals): better parsing for tool calls (#3009)
Fixes #2919

We were assuming a certain structure in `actual_tool_calls` here. A
slight variance when using Gemini broke it. This change makes the
reliability eval work for Gemini without breaking it for the other
models.

Going forward a method like `parse_tool_calls` should exist in the
Gemini model and we should consume tool calls through that method.
Keeping that part of my coming work!

---------

Co-authored-by: ysolanky <yash@phidata.com>
Co-authored-by: Yash Pratap Solanky <101447028+ysolanky@users.noreply.github.com>