fix: Do not leak the full host env vars into MCP tools (#3547)
## Summary

When running MCP with stdio, the environment variables passed to the
server/client should not include the whole environment of the caller, as
it creates a serious security issue.

The `mcp` client implementation already addressed the issue:
https://github.com/modelcontextprotocol/python-sdk/issues/99.

The simplest solution is to just use their utility function for
filtering the environment instead of dumping the whole `os.environ` to
the MCP client.

## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [ ] Code complies with style guidelines
- [x] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [x] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [ ] Tested in clean environment
- [ ] Tests added/updated (if applicable)