Fix: agents deepcopy memory (#3716)
## Summary

Agents/teams that are in teams have to work from a copy of memory. This
is required to ensure consistent management of sessions.
This however does break a few cases for agents/teams that are used
standalone. This change fixes those case.

In that case where teams are used, then memory has to be access from the
actual team/agent instance.

## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [ ] Code complies with style guidelines
- [ ] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [ ] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [ ] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)