[fix] Correctly detect when tool has already been wrapped (#3728)
## Summary

Fixes #3727 

Correctly identify if a tool has already been wrapped to validate its
arguments. The previous implementation was not able to recognize if a
tool was already been wrapped leading to the bug documented in the issue
above.

By adding an attribute to the wrapper we can clearly mark the tool
function as wrapped, allowing to easily recognizing at successive
executions if we should skip wrapping it.


## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [x] Code complies with style guidelines
- [x] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [x] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [x] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)

---------

Co-authored-by: Dirk Brand <dirkbrnd@gmail.com>
Co-authored-by: Dirk Brand <51947788+dirkbrnd@users.noreply.github.com>