[feat] Create a flexible in-memory storage option that can be optionally used with any custom persistence solution (AWS S3, Snowflake, etc.). (#4115)
## Summary

I wanted to utilize Agno but found the storage class offerings to be
limited and supporting one at a time can be a slow process, because of
this I have created an in-memory storage class that can be hooked up to
any custom persistence solution.

I have heavily modeled the implementation after the code for the json
storage class and its tests and cookbook examples.

I have also created a README but I think this would be best served
elsewhere - happy to move it wherever appropriate.

## Type of change

- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [x] Code complies with style guidelines
- [x] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [x] Self-review completed
- [x] Documentation updated (comments, docstrings)
- [x] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [x] Tested in clean environment
- [x] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)

---------

Co-authored-by: Dirk Brand <dirkbrnd@gmail.com>